requirements.txt
asgiref==3.7.2
black==23.12.1
click==8.1.7
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
flake8==7.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.2.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.2.4
